Move nested lib folder out of components
