+++ src/web/components/Card/components/LI.tsx
@@ -3,7 +3,7 @@ import { css, cx } from 'emotion';
 
 import { from, until } from '@guardian/src-foundations/mq';
 
-import { verticalDivider } from '../lib/verticalDivider';
+import { verticalDivider } from '@root/src/web/lib/verticalDivider';
 
 const liStyles = css`
     /* This position relative is needed to contain the veritcal divider */
