+++ src/web/layouts/DecideLayout.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { designTypeDefault } from '@root/src/lib/designTypes';
 
+import { hasShowcase } from '@root/src/web/lib/layoutHelpers';
 import { StandardLayout } from './StandardLayout/StandardLayout';
 import { ShowcaseLayout } from './ShowcaseLayout/ShowcaseLayout';
 import { CommentLayout } from './CommentLayout';
-
-import { hasShowcase } from './layoutHelpers';
 
 type Props = {
     designType: DesignType;
