+++ src/web/lib/decidePillarLight.ts
